Improve cpu utilisation analysis (#104)
* Remove unused dependencies

* Update gopsutil library to v2.19.01

* Simplify and refactor cpu utilization calculation

because the https://github.com/shirou/gopsutil/issues/600 was fixed in the new gopsutil version

* Update cpu Times retrieval and processing for Windows (eliminating slow WMI queries)

Now it uses winapi call to NtQuerySystemInformation.
Added module for wrapping the work with winapi

* Improve listing of top consuming processes for windows

- Use winapi call, where PID is available
- Remove unused functions
- Fix counting of skipped processes for linux top implementation

* Change the way 'Load' value is calculated for processes listed by 'top'

Now the max values is 100%, where 100% is a full load of all available cores.

* Remove 'Load' field from JSON result for the list of 'top' processes

Also, the 'command' field will be reported only if it's not empty. (ex. for Windows it's not available)

* Rename JSON 'Load' fields for the listing of 'top' processes

* Skip process with PID=0 (System Idle) when listing 'top' processes for Windows

* Fix cpu utilisation analysis for the processes with same name

now the PID of the process is available at all systems, so we can use it for comparison

* Added configuration values regarding CPU Utilisation Analysys to result JSON

Also renamed the field for consistency

* Add CPU Utilisation config values to result JSON only if analysis is active