Log location now cagent.log
Previously the line stated that the logs go in to the conf file on the Unix system. I suspect this is unlikely and a simple copy+paste error of the configuration section above.

This change updates the README to describe the true default destination of the logs.

Although still seems weird to write the logs to `/etc` by default?